+++ git-branchless-lib/src/core/mod.rs
@@ -11,3 +11,4 @@ pub mod node_descriptors;
 pub mod repo_ext;
 pub mod rewrite;
 pub mod task;
+pub mod untracked_file_cache;