+++ git-branchless/src/commands/mod.rs
@@ -36,12 +36,14 @@ fn command_main(ctx: CommandContext, opts: Opts) -> EyreExitOr<()> {
         Command::Amend {
             move_options,
             reparent,
+            untracked_file_strategy,
         } => amend::amend(
             &effects,
             &git_run_info,
             &ResolveRevsetOptions::default(),
             &move_options,
             reparent,
+            untracked_file_strategy,
         )?,
 
         Command::BugReport => bug_report::bug_report(&effects, &git_run_info)?,