+++ git-branchless/tests/test_amend.rs
@@ -1,4 +1,8 @@
-use lib::testing::{GitRunOptions, make_git, remove_rebase_lines, trim_lines};
+use lib::testing::{
+    GitRunOptions, make_git,
+    pty::{PtyAction, run_in_pty},
+    remove_rebase_lines, trim_lines,
+};
 
 #[test]
 fn test_amend_with_children() -> eyre::Result<()> {
@@ -387,8 +391,9 @@ fn test_amend_head() -> eyre::Result<()> {
     git.write_file_txt("newfile", "some new file")?;
     {
         let (stdout, _stderr) = git.branchless("amend", &[])?;
-        insta::assert_snapshot!(stdout, @"There are no uncommitted or staged changes. Nothing to amend.
-");
+        insta::assert_snapshot!(stdout, @r###"
+        There are no uncommitted or staged changes. Nothing to amend.
+        "###);
     }
 
     git.run(&["add", "."])?;
@@ -440,6 +445,210 @@ fn test_amend_head_with_file_with_space() -> eyre::Result<()> {
         |
         @ 5ccdda3 create test file with space.txt
         "###);
+    }
+
+    Ok(())
+}
+
+#[test]
+fn test_amend_with_new_untracked_files() -> eyre::Result<()> {
+    //
+    // This test mostly mimics the corresponding test for `record`. Changes here
+    // may also need to be made there.
+    //
+    // See fn test_record_with_new_untracked_files in git-branchless-record/tests/test_record.rs
+    //
+
+    let git = make_git()?;
+    if !git.supports_committer_date_is_author_date()? {
+        return Ok(());
+    }
+    git.init_repo()?;
+    git.detach_head()?;
+    git.commit_file("test1", 1)?;
+
+    {
+        // confirm initial state: only test1
+
+        let (stdout, _stderr) = git.run(&["show", "--pretty=format:", "--stat", "HEAD"])?;
+        insta::assert_snapshot!(&stdout, @"
+            test1.txt | 1 +
+            1 file changed, 1 insertion(+)
+        ");
+    }
+
+    {
+        // working copy & disabled (default) => test2 not added
+        git.write_file_txt("test1", "test1 updated 1")?;
+        git.write_file_txt("test2", "test2 new")?;
+
+        let (stdout, _stderr) = git.branchless("amend", &[])?;
+        insta::assert_snapshot!(stdout, @r###"
+        branchless: running command: <git-executable> reset 7e3c3a90f4d173b07747712f9c8fb7b3bb9902e3 --
+        Amended with 1 uncommitted change.
+        "###);
+
+        let (stdout, _stderr) = git.run(&["show", "--pretty=format:", "--stat", "HEAD"])?;
+        insta::assert_snapshot!(&stdout, @"
+            test1.txt | 1 +
+            1 file changed, 1 insertion(+)
+        ");
+    }
+
+    {
+        // working copy & add
+        git.write_file_txt("test1", "test1 updated 2")?;
+        // test2 should still be considered "new" because last run was disabled
+
+        let (stdout, _stderr) = git.branchless("amend", &["--untracked", "add"])?;
+        insta::assert_snapshot!(stdout, @r###"
+        Including 1 new untracked file: test2.txt
+        branchless: running command: <git-executable> reset c6ba988a5528776b74f8eb23d067b399762358e6 --
+        Amended with 2 uncommitted changes.
+        "###);
+
+        let (stdout, _stderr) = git.run(&["show", "--pretty=format:", "--stat", "HEAD"])?;
+        insta::assert_snapshot!(&stdout, @"
+            test1.txt | 1 +
+            test2.txt | 1 +
+            2 files changed, 2 insertions(+)
+        ");
+    }
+
+    {
+        // working copy & skip
+        git.write_file_txt("test2", "test2 updated 3")?;
+        git.write_file_txt("test3", "test3 new")?;
+
+        let (stdout, _stderr) = git.branchless("amend", &["--untracked", "skip"])?;
+        insta::assert_snapshot!(stdout, @r###"
+        Skipping 1 new untracked file: test3.txt
+        hint: this file will remain skipped and will not be automatically reconsidered
+        hint: to add it yourself: git add
+        hint: disable this hint by running: git config --global branchless.hint.addSkippedFiles false
+        branchless: running command: <git-executable> reset 7fa9d6e9747d464c52ebf08b2477237e9685ff11 --
+        Amended with 1 uncommitted change.
+        "###);
+
+        let (stdout, _stderr) = git.run(&["show", "--pretty=format:", "--stat", "HEAD"])?;
+        insta::assert_snapshot!(&stdout, @"
+            test1.txt | 1 +
+            test2.txt | 1 +
+            2 files changed, 2 insertions(+)
+        ");
+    }
+
+    {
+        // working copy & prompt
+        git.write_file_txt("test2", "test2 updated 4")?;
+        // test3.txt should remain skipped because we've already seen it
+        git.write_file_txt("test4", "test4 new")?;
+
+        run_in_pty(
+            &git,
+            "amend",
+            &["--untracked", "prompt"],
+            &[PtyAction::WaitUntilContains("test4"), PtyAction::Write("y")],
+        )?;
+
+        let (stdout, _stderr) = git.run(&["show", "--pretty=format:", "--stat", "HEAD"])?;
+        insta::assert_snapshot!(&stdout, @"
+            test1.txt | 1 +
+            test2.txt | 1 +
+            test4.txt | 1 +
+            3 files changed, 3 insertions(+)
+        ");
+    }
+
+    {
+        // working copy & add
+        // -> only new files in working copy, no changes to tracked files
+        // test3.txt still skipped
+        git.write_file_txt("test5", "test5 new")?;
+
+        let (stdout, _stderr) = git.branchless("amend", &["--untracked", "add"])?;
+        insta::assert_snapshot!(stdout, @r###"
+        Skipping 1 previously skipped file: test3.txt
+        Including 1 new untracked file: test5.txt
+        branchless: running command: <git-executable> reset dc6825a5feb1d36a022a432bc5f992ead8748a00 --
+        Amended with 1 uncommitted change.
+        "###);
+
+        let (stdout, _stderr) = git.run(&["show", "--pretty=format:", "--stat", "HEAD"])?;
+        insta::assert_snapshot!(&stdout, @"
+            test1.txt | 1 +
+            test2.txt | 1 +
+            test4.txt | 1 +
+            test5.txt | 1 +
+            4 files changed, 4 insertions(+)
+        ");
+    }
+
+    {
+        // index (staged) & add => untracked wc changes ignored
+        git.write_file_txt("test5", "test5 updated\n6")?; // using \n to indicate 2 added lines, below
+        git.write_file_txt("test6", "test6 new")?;
+        git.run(&["add", "test5.txt"])?;
+
+        let (stdout, _stderr) = git.branchless("amend", &["--untracked", "add"])?;
+        insta::assert_snapshot!(stdout, @r###"
+        branchless: running command: <git-executable> reset f734544d45f317a2ce0304ea45c3d62e0e50b1a2 --
+        Amended with 1 staged change.
+        "###);
+
+        let (stdout, _stderr) = git.run(&["show", "--pretty=format:", "--stat", "HEAD"])?;
+        insta::assert_snapshot!(&stdout, @"
+            test1.txt | 1 +
+            test2.txt | 1 +
+            test4.txt | 1 +
+            test5.txt | 2 ++
+            4 files changed, 5 insertions(+)
+        ");
+    }
+
+    {
+        // working copy & disable (again) => still no output about added/skipped files
+        git.write_file_txt("test1", "test1 updated\n7")?; // using \n to indicate 2 added lines, below
+
+        let (stdout, _stderr) = git.branchless("amend", &[])?;
+        insta::assert_snapshot!(stdout, @r###"
+        branchless: running command: <git-executable> reset 93af47a41ecad2de9b2ffe67586df62f6850ace8 --
+        Amended with 1 uncommitted change.
+        "###);
+
+        let (stdout, _stderr) = git.run(&["show", "--pretty=format:", "--stat", "HEAD"])?;
+        insta::assert_snapshot!(&stdout, @"
+            test1.txt | 2 ++
+            test2.txt | 1 +
+            test4.txt | 1 +
+            test5.txt | 2 ++
+            4 files changed, 6 insertions(+)
+        ");
+    }
+
+    {
+        // working copy & add (again)
+        //  - test3 still skipped
+        //  - test6 added because it was totally skipped during last/disabled run
+        git.write_file_txt("test1", "test1 updated 8")?;
+
+        let (stdout, _stderr) = git.branchless("amend", &["--untracked", "add"])?;
+        insta::assert_snapshot!(stdout, @r###"
+        Skipping 1 previously skipped file: test3.txt
+        Including 1 new untracked file: test6.txt
+        branchless: running command: <git-executable> reset d52a98c726910a75f2b9163665d849236c918442 --
+        Amended with 2 uncommitted changes.
+        "###);
+
+        let (stdout, _stderr) = git.run(&["show", "--pretty=format:", "--stat", "HEAD"])?;
+        insta::assert_snapshot!(&stdout, @"
+            test1.txt | 1 +
+            test2.txt | 1 +
+            test4.txt | 1 +
+            test5.txt | 2 ++
+            test6.txt | 1 +
+            5 files changed, 6 insertions(+)
+        ");
     }
 
     Ok(())