+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - (#1461): added `!` revset postfix operator as shortcut for "only child"
 - (#1464): created `git split` command to extract changes from a commit
 - (#1603): added `git move --dry-run` to test in-memory rebases
+- (#1604): `git record` and `git amend` can now automatically detect and begin tracking new files (optional, disabled by default)
 - (#1632): added `git record --fixup` option, to create a fixup commit (similar to `reword --fixup`)
 
 ### Changed